+++ src/commands/error/log/view.ts
@@ -1,67 +1,10 @@
-/**
- * Copyright (C) 2021 PythonCoderAS
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-import { SlashCommandBuilder } from "@discordjs/builders";
 import { CommandInteraction, MessageEmbed } from "discord.js";
 import { DateTime } from "luxon";
 import { Op } from "sequelize";
-import { ErrorLog } from "../orm/errorLog";
+import { ErrorLog } from "../../../orm/errorLog";
 
 module.exports = {
-  data: new SlashCommandBuilder()
-    .setName("view_error_log")
-    .setDescription("View the bot's error log.")
-    .addIntegerOption((option) =>
-      option
-        .setName("before")
-        .setDescription("Show the errors before this error #")
-        .setRequired(false)
-    )
-    .addIntegerOption((option) =>
-      option
-        .setName("after")
-        .setDescription("Show the errors after this error #")
-        .setRequired(false)
-    )
-    .addIntegerOption((option) =>
-      option
-        .setName("after_time")
-        .setDescription("Show errors after the given UNIX timestamp")
-        .setRequired(false)
-    )
-    .addIntegerOption((option) =>
-      option
-        .setName("before_time")
-        .setDescription("Show errors before the given UNIX timestamp")
-        .setRequired(false)
-    ).addBooleanOption((option) =>
-      option
-        .setName("desc")
-        .setDescription("Show the errors in descending order, default true")
-        .setRequired(false)
-    ).addBooleanOption((option) =>
-      option
-        .setName("with_github_issue_number")
-        .setDescription("Show the errors with github issue numbers, default false")
-        .setRequired(false)
-    ).addStringOption((option) =>
-      option
-        .setName("type_starts_with")
-        .setDescription("Shows the errors with a type starting with the given string")
-        .setRequired(false)),
+  name: "view",
   async execute(interaction: CommandInteraction) {
     const isDesc = interaction.options.getBoolean("desc", false) || true;
     const whereQuery: { [k: string]: object } = {};
@@ -177,4 +120,4 @@ module.exports = {
     embed.addField("Search Properties", fieldData);
     return await interaction.reply({ embeds: [embed] });
   }
-};
+}
